third_party/blink/web_tests/http/tests/loading/onload-vs-immediate-refresh-expected.txt

main frame - DidStartNavigation
main frame - ReadyToCommitNavigation
main frame - didCommitLoadForFrame
main frame - didReceiveTitle: 
main frame - didFinishDocumentLoadForFrame
ALERT: SUCCESS
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
main frame - DidStartNavigation
main frame - ReadyToCommitNavigation
main frame - didCommitLoadForFrame
main frame - didReceiveTitle: 
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame
main frame - didFinishLoadForFrame
You should have seen an alert.
